Struct sos_intrusive::list::List

pub struct List<T, N> where
    T: OwnedRef<N>,
    N: Node,  { /* fields omitted */ }

The List struct is our way of interacting with an intrusive list.

It stores a pointer to the head and tail of the list, the length of the list, and a PhantomData marker for the list's OwnedRef type. It provides the methods for pushing, popping, and indexing the list.

Methods

impl<T, N> List<T, N> where
T: OwnedRef<N>,
N: Node,

const fn new() -> Self

Construct a new List<T, N> with zero elements

fn len(&self) -> usize

Returns the length of the list

fn front(&self) -> Option<&N>

Borrows the first element of the list as an Option

Returns

Some(&N) if the list has elements

None if the list is empty.

fn back(&self) -> Option<&N>

Borrows the last element of the list as an Option

Returns

Some(&N) if the list has elements

None if the list is empty.

fn front_mut(&mut self) -> Option<&mut N>

Mutably borrows the first element of the list as an Option

Returns

Some(&mut N) if the list has elements

None if the list is empty.

fn back_mut(&mut self) -> Option<&mut N>

Mutably borrows the last element of the list as an Option

Returns

Some(&mut N) if the list has elements

None if the list is empty.

fn is_empty(&self) -> bool

Returns true if the list is empty.

fn push_front(&mut self, item: T)

Push an element to the front of the list.

fn push_back(&mut self, item: T)

Push an element to the back of the list.

fn pop_front(&mut self) -> Option<T>

Removes and returns the element at the front of the list.

Returns

Some(T) containing the element at the front of the list if the list is not empty

None if the list is empty

fn pop_back(&mut self) -> Option<T>

Removes and returns the element at the end of the list.

Returns

Some(T) containing the element at the end of the list if the list is not empty

None if the list is empty

fn peek_front(&self) -> Option<&N>

Borrows the element at the front of the list

Returns

Some(&T) containing the element at the end of the list if the list is not empty

None if the list is empty

fn cursor_mut<'a>(&'a mut self) -> ListCursorMut<'a, T, N>

Returns a cursor for iterating over or modifying the list.