Struct sos_intrusive::list::ListCursorMut

pub struct ListCursorMut<'a, T, N> where
    T: OwnedRef<N>,
    T: 'a,
    N: Node,
    N: 'a,  { /* fields omitted */ }

A cursor for an intrusive linked list.

A cursor functions similarly to an iterator, except that it can seek back and forth rather than just advancing through the list, and it can mutate the element "under" the cursor.

A cursor begins before the first element in the list, and once it has been advanced past the last element of the list, it "loops around" back to the first element.

Methods

impl<'a, T, N> ListCursorMut<'a, T, N> where
T: OwnedRef<N>,
T: 'a,
N: Node,
N: 'a,

fn next(&mut self) -> Option<&mut N>

Advances the cursor to the next element and borrows it mutably.

If the cursor is at the end of the list, this advances it back to the first element.

Returns

Some(&mut N) if the list is not empty

None if the list is empty

fn prev(&mut self) -> Option<&mut N>

Steps back the cursor to the previous element and borrows it mutably.

Returns

Some(&mut N) if the list is not empty

None if the list is empty

fn peek_next(&self) -> Option<&N>

Borrows the next element in the list without advancing the cursor.

If the cursor is at the end of the list, this returns the first element instead.

Returns

Some(&N) if the list is not empty

None if the list is empty

fn peek_prev(&self) -> Option<&N>

Borrows the previous element without stepping back the cursor.

If the cursor is at the head of the list, this returns None.

Returns

Some(&N) if the list is not empty

None if the list is empty or if the cursor is at the head of the list

fn remove(&mut self) -> Option<T>

Removes the element currently under the cursor and returns it.

Returns

Some(T) if the there is an element currently under the cursor (i.e., the list is not empty)

None if the list is empty.

fn find_and_remove<P>(&mut self, predicate: P) -> Option<T> where
P: Fn(&N) -> bool,

Searches for and removes the first element matching a predicate.

Arguments

p: A predicate (function of the form &N -> bool) that returns true if the element should be removed and false if it should not.

Returns

Some(T) if an element matching the predicate was found

None if no elements matched the predicate (or if the list is empty.)

fn seek_forward(&mut self, n: usize) -> Option<&mut N>

Advances the cursor n elements and mutably borrows the final element.

This will wrap the cursor around the list if n > the length of the list.

fn seek_backward(&mut self, n: usize) -> Option<&mut N>

Moves the cursor back n times and mutably borrows the final element.

This will wrap the cursor around the list if n > the length of the list.