Trait sos_intrusive::list::Node

pub trait Node: Sized {
    fn next(&self) -> &RawLink<Self>;
    fn prev(&self) -> &RawLink<Self>;
    fn next_mut(&mut self) -> &mut RawLink<Self>;
    fn prev_mut(&mut self) -> &mut RawLink<Self>;
}

This trait defines a node in an intrusive list.

A Node must be capable of providing mutable and immutable references to the previous and next nodes in the list.

Required Methods

fn next(&self) -> &RawLink<Self>

fn prev(&self) -> &RawLink<Self>

fn next_mut(&mut self) -> &mut RawLink<Self>

fn prev_mut(&mut self) -> &mut RawLink<Self>